+++ release_notes.md
@@ -6,6 +6,7 @@ Spider Changelog
 
 ### New Feature
 * Expression now support larger range of function. 
+* Add extern memory address for static allocators.
 
 ### Changes
 * Expression parser is more robust and less prone to errors.